+++ Chocolatey/OfficeDevPnPCmdlets16/tools/chocolateyUninstall.ps1
@@ -1,2 +1,2 @@
 ﻿$name = "officedevpnppowershellcommands16"
-Uninstall-ChocolateyPackage $name 'msi' "https://github.com/OfficeDev/PnP-PowerShell/releases/download/v1.2/PnPPowerShellCommands16.msi /qn"
+Uninstall-ChocolateyPackage $name 'msi' "https://github.com/OfficeDev/PnP-PowerShell/releases/download/v1.3.2/PnPPowerShellCommands16.msi /qn"