+++ changelog.md
@@ -1,5 +1,22 @@
 # OfficeDevPnP.PowerShell Changelog #
 
+**2015-09-02**
+
+* Started adding unit tests
+* Added warning when using Install-SPOSolution to documentation. The cmdlet can potentially clear the composed look gallery.
+
+**2015-08-18**
+
+* Added Set-SPOTraceLog cmdlet
+
+**2015-08-15**
+
+* Added -Recurse parameter to Get-SPOSubWebs cmdlet to recursively retrieve all subwebs
+
+**2015-08-14**
+
+* Modified Connect-SPOnline to output version number when specifying -Verbose parameter
+
 **2015-08-10**
 
 * Added Get-SPOWebPartXml cmdlet to export webpart XML from a page.